+++ packages/forge-proto/index.d.ts
@@ -119,7 +119,9 @@ declare namespace Enums {
     ROLE_ASSET = 6,
     ROLE_STAKE = 7,
     ROLE_VALIDATOR = 8,
-    ROLE_TX = 9,
+    ROLE_GROUP = 9,
+    ROLE_TX = 10,
+    ROLE_ANY = 63,
   }
 
   export enum UpgradeType {
@@ -319,7 +321,9 @@ declare namespace Messages {
     6 = ROLE_ASSET,
     7 = ROLE_STAKE,
     8 = ROLE_VALIDATOR,
-    9 = ROLE_TX,
+    9 = ROLE_GROUP,
+    10 = ROLE_TX,
+    63 = ROLE_ANY,
   }
 
   export enum UpgradeType {